chore: api 함수 변경
