+++ src/apis/index.ts
@@ -1,5 +1,9 @@
 import { authApis } from './auth';
+import { gatheringApis } from './gathering';
+import { placeApis } from './place';
 
 export const apis = {
   auth: authApis,
+  gathering: gatheringApis,
+  place: placeApis,
 } as const;
